refactor(web): move app shell, header and auth guard into route-group layouts
Split src/app into (app) and (auth) route groups. (app)/layout.tsx mounts
AuthenticatedUserProvider — the single useAuthenticatedUser() instance, the
single "not signed in yet" guard screen, and the session shared through
useAuthenticatedUserContext() — around AppShell (background, AppHeader,
content container). (auth)/layout.tsx mounts AuthShell (same background,
brand-only BrandHeader, card slot). Pages now render their content only: no
header, no gradient, no container, no per-page auth spinner.

Because a layout is not remounted on navigation inside its group, the
stored-token check in useAuthenticatedUser() is keyed to usePathname() so an
expired JWT still redirects to /login mid-session; it reuses the previous
user object when the email is unchanged, so re-validating never refetches
GET /users/me.
