+++ apps/web/src/app/(app)/meetings/[id]/page.tsx
@@ -3,7 +3,6 @@
 import { useCallback, useEffect, useRef, useState } from 'react';
 import { useParams, useRouter } from 'next/navigation';
 import { Button, Card, Spinner } from '@heroui/react';
-import { useAuthenticatedUser } from '@/lib/useAuthenticatedUser';
 import {
   ApiError,
   getMeeting,
@@ -11,7 +10,6 @@ import {
   type Recording,
 } from '@/lib/api';
 import { formatDateTime } from '@/lib/format';
-import { AppHeader } from '@/components/layout/AppHeader';
 import { ArrowLeftIcon, CalendarIcon, UsersIcon } from '@/components/icons';
 import { MeetingSummarySection } from '@/components/meetings/MeetingSummarySection';
 import { RecordingCard } from '@/components/meetings/RecordingCard';
@@ -31,7 +29,6 @@ export default function MeetingDetailPage() {
   const meetingId = params.id;
   const router = useRouter();
 
-  const { user, signOut } = useAuthenticatedUser();
   const [meeting, setMeeting] = useState<MeetingDetail | null>(null);
   const [error, setError] = useState<string | null>(null);
   // Bumped by any direct local update (upload/delete) so a poll response already
@@ -39,10 +36,6 @@ export default function MeetingDetailPage() {
   const pollGenerationRef = useRef(0);
 
   useEffect(() => {
-    if (!user) {
-      return;
-    }
-
     getMeeting(meetingId)
       .then(setMeeting)
       .catch((err: unknown) => {
@@ -52,7 +45,7 @@ export default function MeetingDetailPage() {
             : 'Could not load the meeting. Please try again.',
         );
       });
-  }, [meetingId, user]);
+  }, [meetingId]);
 
   const hasPendingRecording =
     meeting?.recordings.some(
@@ -114,7 +107,7 @@ export default function MeetingDetailPage() {
     meeting !== null && recordingsSignature !== reconciledRecordingsSignature;
 
   useEffect(() => {
-    if (!user || (!hasPendingRecording && !isSummaryPending)) {
+    if (!hasPendingRecording && !isSummaryPending) {
       return;
     }
 
@@ -140,7 +133,7 @@ export default function MeetingDetailPage() {
     }, 4000);
 
     return () => clearInterval(intervalId);
-  }, [meetingId, user, hasPendingRecording, isSummaryPending]);
+  }, [meetingId, hasPendingRecording, isSummaryPending]);
 
   const handleUploaded = useCallback((recording: Recording) => {
     pollGenerationRef.current += 1;
@@ -172,102 +165,83 @@ export default function MeetingDetailPage() {
     );
   }, []);
 
-  if (!user) {
-    return (
-      <div className="flex min-h-screen items-center justify-center">
-        <Spinner aria-label="Loading" />
-      </div>
-    );
-  }
-
   return (
-    <div className="flex min-h-screen flex-col bg-linear-to-br from-accent/10 via-background to-background">
-      <AppHeader email={user.email} onSignOut={signOut} />
-      <div className="flex justify-center px-4 py-12">
-        <div className="flex w-full max-w-2xl flex-col gap-6">
-          <Button
-            className="h-11 self-start md:h-10"
-            variant="secondary"
-            onPress={() => router.back()}
-          >
-            <ArrowLeftIcon className="size-4" />
-            Back
-          </Button>
-          {error ? (
-            <Card>
-              <Card.Content>
-                <p className="text-sm text-danger" role="alert">
-                  {error}
+    <>
+      <Button
+        className="h-11 self-start md:h-10"
+        variant="secondary"
+        onPress={() => router.back()}
+      >
+        <ArrowLeftIcon className="size-4" />
+        Back
+      </Button>
+      {error ? (
+        <Card>
+          <Card.Content>
+            <p className="text-sm text-danger" role="alert">
+              {error}
+            </p>
+          </Card.Content>
+        </Card>
+      ) : meeting === null ? (
+        <div className="flex justify-center py-12">
+          <Spinner aria-label="Loading meeting" />
+        </div>
+      ) : (
+        <>
+          <Card>
+            <Card.Header>
+              <Card.Title>{meeting.title}</Card.Title>
+            </Card.Header>
+            <Card.Content className="flex flex-col gap-1.5">
+              <p className="flex items-center gap-1.5 text-sm text-muted">
+                <CalendarIcon aria-hidden="true" className="size-4 shrink-0" />
+                {formatDateTime(meeting.date)}
+              </p>
+              {meeting.participants.length > 0 ? (
+                <p className="flex items-center gap-1.5 text-sm text-muted">
+                  <UsersIcon aria-hidden="true" className="size-4 shrink-0" />
+                  {meeting.participants.join(', ')}
                 </p>
+              ) : null}
+            </Card.Content>
+          </Card>
+
+          <Card>
+            <Card.Header>
+              <Card.Title>Recordings</Card.Title>
+            </Card.Header>
+            <Card.Content className="flex flex-col gap-6">
+              <RecordingUploader
+                meetingId={meeting.id}
+                onUploaded={handleUploaded}
+              />
+              {meeting.recordings.map((recording) => (
+                <RecordingCard
+                  key={recording.id}
+                  meetingId={meeting.id}
+                  onDeleted={handleDeleted}
+                  recording={recording}
+                />
+              ))}
+            </Card.Content>
+          </Card>
+
+          {showSummarySection ? (
+            <Card>
+              <Card.Header>
+                <Card.Title>Summary</Card.Title>
+              </Card.Header>
+              <Card.Content>
+                <MeetingSummarySection
+                  isUpdating={isSummaryPending}
+                  summary={meeting.summary}
+                />
               </Card.Content>
             </Card>
-          ) : meeting === null ? (
-            <div className="flex justify-center py-12">
-              <Spinner aria-label="Loading meeting" />
-            </div>
-          ) : (
-            <>
-              <Card>
-                <Card.Header>
-                  <Card.Title>{meeting.title}</Card.Title>
-                </Card.Header>
-                <Card.Content className="flex flex-col gap-1.5">
-                  <p className="flex items-center gap-1.5 text-sm text-muted">
-                    <CalendarIcon
-                      aria-hidden="true"
-                      className="size-4 shrink-0"
-                    />
-                    {formatDateTime(meeting.date)}
-                  </p>
-                  {meeting.participants.length > 0 ? (
-                    <p className="flex items-center gap-1.5 text-sm text-muted">
-                      <UsersIcon
-                        aria-hidden="true"
-                        className="size-4 shrink-0"
-                      />
-                      {meeting.participants.join(', ')}
-                    </p>
-                  ) : null}
-                </Card.Content>
-              </Card>
-
-              <Card>
-                <Card.Header>
-                  <Card.Title>Recordings</Card.Title>
-                </Card.Header>
-                <Card.Content className="flex flex-col gap-6">
-                  <RecordingUploader
-                    meetingId={meeting.id}
-                    onUploaded={handleUploaded}
-                  />
-                  {meeting.recordings.map((recording) => (
-                    <RecordingCard
-                      key={recording.id}
-                      meetingId={meeting.id}
-                      onDeleted={handleDeleted}
-                      recording={recording}
-                    />
-                  ))}
-                </Card.Content>
-              </Card>
-
-              {showSummarySection ? (
-                <Card>
-                  <Card.Header>
-                    <Card.Title>Summary</Card.Title>
-                  </Card.Header>
-                  <Card.Content>
-                    <MeetingSummarySection
-                      isUpdating={isSummaryPending}
-                      summary={meeting.summary}
-                    />
-                  </Card.Content>
-                </Card>
-              ) : null}
-            </>
-          )}
-        </div>
-      </div>
-    </div>
+          ) : null}
+        </>
+      )}
+    </>
   );
 }
