+++ apps/web/src/components/layout/AppHeader.tsx
@@ -1,15 +1,20 @@
 import { Button, Link } from '@heroui/react';
-import { VideoCameraIcon } from '@/components/icons';
+import { BrandHeader } from '@/components/layout/BrandHeader';
 import { UserAvatar } from '@/components/profile/UserAvatar';
 
 interface AppHeaderProps {
-  email?: string;
+  email: string;
   name?: string | null;
   hasAvatar?: boolean;
   avatarUpdatedAt?: string | null;
-  onSignOut?: () => void;
+  onSignOut: () => void;
 }
 
+/**
+ * The header of every authenticated route: the brand bar plus the signed-in user's
+ * avatar/name (linking to `/profile`) and a sign-out control. Rendered once, by
+ * `AppShell`; a page that renders it itself is a bug.
+ */
 export function AppHeader({
   email,
   name = null,
@@ -18,43 +23,22 @@ export function AppHeader({
   onSignOut,
 }: AppHeaderProps) {
   return (
-    <header className="flex items-center justify-between border-b border-default-200 bg-background px-4 py-3 sm:px-6">
-      <Link className="flex items-center gap-2 rounded-lg py-1" href="/">
-        <span className="flex size-9 shrink-0 items-center justify-center rounded-lg bg-accent text-accent-foreground">
-          <VideoCameraIcon aria-hidden="true" className="size-5" />
+    <BrandHeader>
+      <Link className="flex items-center gap-2 rounded-lg py-1" href="/profile">
+        <UserAvatar
+          avatarUpdatedAt={avatarUpdatedAt}
+          email={email}
+          hasAvatar={hasAvatar}
+          name={name}
+          size="header"
+        />
+        <span className="text-xs leading-tight text-muted">
+          {name?.trim() || email}
         </span>
-        <h1 className="text-base font-semibold leading-tight">
-          Video Meetings
-        </h1>
       </Link>
-      <div className="flex items-center gap-3">
-        {email ? (
-          <Link
-            className="flex items-center gap-2 rounded-lg py-1"
-            href="/profile"
-          >
-            <UserAvatar
-              avatarUpdatedAt={avatarUpdatedAt}
-              email={email}
-              hasAvatar={hasAvatar}
-              name={name}
-              size="header"
-            />
-            <span className="text-xs leading-tight text-muted">
-              {name?.trim() || email}
-            </span>
-          </Link>
-        ) : null}
-        {onSignOut ? (
-          <Button
-            className="h-11 md:h-10"
-            variant="secondary"
-            onPress={onSignOut}
-          >
-            Sign out
-          </Button>
-        ) : null}
-      </div>
-    </header>
+      <Button className="h-11 md:h-10" variant="secondary" onPress={onSignOut}>
+        Sign out
+      </Button>
+    </BrandHeader>
   );
 }
